Fixed all known WaitingRoom's bugs.
To login to WaitingRoom you HAVE TO call get_waiting_room.
To logout from WaitingRoom call del_waiting_room.

Logging in/out to/from WaitingRoom is not the same as
signing in/out as a user (see Session methods)

You can set callbacks to waiting room to have frontend updated
on status changes of users.